requirements.txt
markdownify
beautifulsoup4
PyPDF2
lxml
requests
openagents-node-sdk
python-dateutil